# backend/internal/interface/api/rest/api-specs/openapi/scout/openapi.yaml
openapi: 3.0.3
info:
  title: Scout API
  version: 2.0.0
  description: |
    Scout is a greenhouse pest & disease monitoring service. Cameras photograph
    the crop and an ML model detects pests/diseases (bounding boxes). Photo
    metadata and predictions are read from a SQLite dataset; the original images
    are ingested into object storage (MinIO) through a presigned-upload flow.

    ## Auth
    JWT bearer auth. Obtain a token from `POST /api/v1/auth/login`, then send it
    as `Authorization: Bearer <token>`. Image endpoints (`/thumbnail`, `/original`)
    additionally accept the token as a `?token=` query parameter so they can be
    used directly in `<img>` tags.

    Roles: `user` can browse photos; `admin` can additionally create upload links,
    check object existence, and read `/metrics` and `/debug/vars`.

    ## Bounding boxes
    Normalized to `[0,1]` against the original image. Two corners: (xMin,yMin)
    top-left and (xMax,yMax) bottom-right. The frontend multiplies these by the
    ACTUAL rendered image size.

    ## Errors
    Every error response shares `request_id` and `message`, plus a machine-readable
    `code`. Errors are shaped in one place; no stack traces are leaked.
servers:
  - url: http://localhost:8080
    description: Local backend (Docker Compose)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/photos/{photoId}/upload-link:
    post:
      tags: [Photos]
      operationId: createUploadLink
      summary: Get a presigned PUT URL to upload the original bytes (admin only)
      parameters:
        - $ref: '#/components/parameters/PhotoId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UploadLinkRequest' }
      responses:
        '200':
          description: Presigned upload link (includes `exists` for idempotency)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadLink' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PhotoId:
      name: photoId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: The request was invalid.
      content:
        application/json: { schema: { $ref: '#/components/schemas/ValidationError' } }
    Unauthorized:
      description: Missing or invalid token.
      content:
        application/json: { schema: { $ref: '#/components/schemas/ApiError' } }
    Forbidden:
      description: Insufficient role.
      content:
        application/json: { schema: { $ref: '#/components/schemas/ApiError' } }
    NotFound:
      description: Resource not found.
      content:
        application/json: { schema: { $ref: '#/components/schemas/NotFoundError' } }
  schemas:
...
    UploadLinkRequest:
      type: object
      required: [contentType]
      properties:
        contentType: { type: string, example: image/jpeg }
    UploadLink:
      type: object
      required: [url, method, expiresAt, exists]
      properties:
        url: { type: string, format: uri }
        method: { type: string, enum: [PUT] }
        headers:
          type: object
          additionalProperties: { type: string }
        expiresAt: { type: string, format: date-time }
        exists: { type: boolean, description: Whether the object already exists. }
...
    ApiError:
      type: object
      required: [request_id, message]
      properties:
        request_id: { type: string }
        message: { type: string }
        code: { type: string }
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/ApiError'
      properties:
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              issue: { type: string }
    NotFoundError:
      allOf:
        - $ref: '#/components/schemas/ApiError'
      properties:
        resource_id: { type: string }